+++ mayan/apps/sources/storages.py
@@ -9,16 +9,13 @@ from .settings import (
 )
 
 storage_source_cache = DefinedStorage(
-    dotted_path=setting_source_cache_storage_backend.value,
-    error_message=_(
+    dotted_path=setting_source_cache_storage_backend.value, error_message=_(
         message='Unable to initialize the staging folder file image '
         'storage. Check the settings {} and {} for formatting '
         'errors.'.format(
             setting_source_cache_storage_backend.global_name,
             setting_source_cache_storage_backend_arguments.global_name
         )
-    ),
-    label=_(message='Staging folder files'),
-    name=STORAGE_NAME_SOURCE_CACHE_FOLDER,
-    kwargs=setting_source_cache_storage_backend_arguments.value
+    ), kwargs=setting_source_cache_storage_backend_arguments.value,
+    label=_(message='Sources cache'), name=STORAGE_NAME_SOURCE_CACHE_FOLDER
 )